Improve key handle / keypress / password support
Previously, the U2F key handle was preselected and sent to openvpn,
before we attempted to get a keypress from the U2F device.

This caused two problems:

- You would get a "please enter your password" GUI prompt before you had
  done anything: sending the key handle in the username field already
  required permissions.

- Having the choice between multiple different key handles locally would
  not be possible.

Now, the key handle and challenge/response are generated at
username-request time. The response is stored for later when openvpn
asks for the password.

The main advantage of this is that you don't get a password prompt out
of the blue. (And multiple key handles should now be possible.)

The disadvantage is that the challenge has more time to go stale when
you have to fill out your password _after_ the U2F keypress. Because of
this, the valid-challege-time has been increased from 8 seconds to 20
seconds on the server. That should be more than sufficient to fill out
your password. (And if it isn't, openvpn will retry the authentication
quickly anyway, and your system-password will be cached by then.)